trying a collision system

+++ include/utils.hpp
@@ -81,4 +81,4 @@ math initMath();
 float vectorCrossProduct(Vector2, Vector2);
 bool rangesOverlap(float, float, float, float);
 bool boxesOverlap(Vector2, Vector2, Vector2, Vector2);
-int sideLine(Vector2, Vector2, Vector2);
+float sideLine(Vector2, Vector2, Vector2);
